add a (very limited) graphql backend
The SQLAlchemy code is moved into a separate blueprint
